Renamed 'values' folder in 'tree' folder  to 'structure'.
