+++ spec/tree/builder/buildTreeSpec.js
@@ -1,14 +1,14 @@
 import buildTree from "../../../src/tree/builder/buildTree";
 import treeToStringifiedJSON from "../../../src/tree/utils/treeToStringifiedJSON";
-import FormattingElement from "../../../src/tree/values/FormattingElement";
-import Heading from "../../../src/tree/values/nodes/Heading";
-import List from "../../../src/tree/values/nodes/List";
-import ListItem from "../../../src/tree/values/nodes/ListItem";
-import Paragraph from "../../../src/tree/values/nodes/Paragraph";
+import FormattingElement from "../../../src/tree/structure/FormattingElement";
+import Heading from "../../../src/tree/structure/nodes/Heading";
+import List from "../../../src/tree/structure/nodes/List";
+import ListItem from "../../../src/tree/structure/nodes/ListItem";
+import Paragraph from "../../../src/tree/structure/nodes/Paragraph";
 
-import StructuredIrrelevant from "../../../src/tree/values/nodes/StructuredIrrelevant";
-import StructuredNode from "../../../src/tree/values/nodes/StructuredNode";
-import TextContainer from "../../../src/tree/values/nodes/TextContainer";
+import StructuredIrrelevant from "../../../src/tree/structure/nodes/StructuredIrrelevant";
+import StructuredNode from "../../../src/tree/structure/nodes/StructuredNode";
+import TextContainer from "../../../src/tree/structure/nodes/TextContainer";
 
 import htmlFile from "../../fullTextTests/testTexts/en/englishPaper1.html";
 import htmlFile2 from "../../fullTextTests/testTexts/de/germanPaper2.html";
