+++ spec/tree/structure/nodes/ParagraphSpec.js
@@ -1,4 +1,4 @@
-import Paragraph from "../../../../src/tree/values/nodes/Paragraph";
+import Paragraph from "../../../../src/tree/structure/nodes/Paragraph";
 
 describe( "Paragraph tree node", () => {
 	describe( "constructor", () => {
