+++ src/tree/structure/FormattingElement.js
