+++ src/tree/structure/nodes/Node.js
