+++ src/tree/structure/nodes/StructuredIrrelevant.js
