+++ src/tree/structure/nodes/StructuredNode.js
