+++ src/tree/structure/nodes/TextContainer.js
