+++ src/tree/structure/nodes/Whitespace.js
